Remove dependency on non-existing ppx_deriving (#336)
When the ppx_deriving package is vendored and not installed with dune
the library "ppx_deriving" can't be located. ppx_deriving contains
several libraries under the "ppx_deriving." prefix, but not library
simply named "ppx_deriving". This works when ppx_deriving is installed
with opam due to inconsistencies with how dune locates libraries between
vendored and opam-installed packages.